Optimize refactor
