+++ packages/manager/src/features/Kubernetes/KubernetesPlansPanel/KubernetesPlanContainer.tsx
@@ -3,16 +3,11 @@ import Grid from '@mui/material/Unstable_Grid2';
 import * as React from 'react';
 
 import { Hidden } from 'src/components/Hidden';
-import { Table } from 'src/components/Table';
-import { TableBody } from 'src/components/TableBody';
-import { TableCell } from 'src/components/TableCell';
-import { TableHead } from 'src/components/TableHead';
-import { TableRow } from 'src/components/TableRow';
-import { TableRowEmpty } from 'src/components/TableRowEmpty/TableRowEmpty';
 import { useFlags } from 'src/hooks/useFlags';
 import { PLAN_SELECTION_NO_REGION_SELECTED_MESSAGE } from 'src/utilities/pricing/constants';
 
 import { KubernetesPlanSelection } from './KubernetesPlanSelection';
+import { KubernetesPlanSelectionTable } from './KubernetesPlanSelectionTable';
 
 import type { LinodeTypeClass } from '@linode/api-v4';
 import type {
@@ -20,16 +15,6 @@ import type {
   PlanSelectionFilterOptionsTable,
 } from 'src/features/components/PlansPanel/PlanContainer';
 import type { PlanWithAvailability } from 'src/features/components/PlansPanel/types';
-
-const tableCells = [
-  { cellName: 'Plan', center: false, noWrap: false, testId: 'plan' },
-  { cellName: 'Monthly', center: false, noWrap: false, testId: 'monthly' },
-  { cellName: 'Hourly', center: false, noWrap: false, testId: 'hourly' },
-  { cellName: 'RAM', center: true, noWrap: false, testId: 'ram' },
-  { cellName: 'CPUs', center: true, noWrap: false, testId: 'cpu' },
-  { cellName: 'Storage', center: true, noWrap: false, testId: 'storage' },
-  { cellName: 'Quantity', center: false, noWrap: false, testId: 'quantity' },
-];
 
 export interface KubernetesPlanContainerProps {
   allDisabledPlans: PlanWithAvailability[];
@@ -162,99 +147,36 @@ export const KubernetesPlanContainer = (
           {planSelectionDividers.map((planSelectionDivider) =>
             planType === planSelectionDivider.planType &&
             planSelectionDivider.flag ? (
-              planSelectionDivider.tables.map((table) => {
+              planSelectionDivider.tables.map((table, idx) => {
                 const filteredPlans = table.planFilter
                   ? plans.filter(table.planFilter)
                   : plans;
                 return (
                   filteredPlans.length > 0 && (
-                    <Table
-                      aria-label={`List of ${table.header} Plans`}
-                      key={table.header}
-                      spacingBottom={16}
-                    >
-                      <TableHead>
-                        <TableRow>
-                          {tableCells.map(
-                            ({ cellName, center, noWrap, testId }) => {
-                              const attributeValue = `${testId}-header`;
-                              return (
-                                <TableCell
-                                  center={center}
-                                  data-qa={attributeValue}
-                                  key={testId}
-                                  noWrap={noWrap}
-                                >
-                                  {cellName === 'Quantity' ? (
-                                    <p className="visually-hidden">
-                                      {cellName}
-                                    </p>
-                                  ) : cellName === 'Plan' ? (
-                                    table.header
-                                  ) : (
-                                    cellName
-                                  )}
-                                </TableCell>
-                              );
-                            }
-                          )}
-                        </TableRow>
-                      </TableHead>
-                      <TableBody role="grid">
-                        {shouldDisplayNoRegionSelectedMessage ? (
-                          <TableRowEmpty
-                            colSpan={9}
-                            message={PLAN_SELECTION_NO_REGION_SELECTED_MESSAGE}
-                          />
-                        ) : (
-                          renderPlanSelection({
-                            header: table.header,
-                            planFilter: table.planFilter,
-                          })
-                        )}
-                      </TableBody>
-                    </Table>
+                    <KubernetesPlanSelectionTable
+                      renderPlanSelection={() =>
+                        renderPlanSelection({
+                          header: table.header,
+                          planFilter: table.planFilter,
+                        })
+                      }
+                      shouldDisplayNoRegionSelectedMessage={
+                        shouldDisplayNoRegionSelectedMessage
+                      }
+                      filterOptions={table}
+                      key={`k8-plan-filter-${idx}`}
+                    />
                   )
                 );
               })
             ) : (
-              <Table
-                aria-label="List of Linode Plans"
+              <KubernetesPlanSelectionTable
+                shouldDisplayNoRegionSelectedMessage={
+                  shouldDisplayNoRegionSelectedMessage
+                }
                 key={planType}
-                spacingBottom={16}
-              >
-                <TableHead>
-                  <TableRow>
-                    {tableCells.map(({ cellName, center, noWrap, testId }) => {
-                      const attributeValue = `${testId}-header`;
-                      return (
-                        <TableCell
-                          center={center}
-                          data-qa={attributeValue}
-                          key={testId}
-                          noWrap={noWrap}
-                        >
-                          {cellName === 'Quantity' ? (
-                            <p className="visually-hidden">{cellName}</p>
-                          ) : (
-                            cellName
-                          )}
-                        </TableCell>
-                      );
-                    })}
-                  </TableRow>
-                </TableHead>
-                <TableBody role="grid">
-                  {shouldDisplayNoRegionSelectedMessage ? (
-                    <TableRowEmpty
-                      colSpan={9}
-                      message={PLAN_SELECTION_NO_REGION_SELECTED_MESSAGE}
-                    />
-                  ) : (
-                    renderPlanSelection()
-                  )}
-                </TableBody>
-              </Table>
+                renderPlanSelection={renderPlanSelection}
+              />
             )
           )}
         </Grid>
